consistent offer pdf naming (#510)
the naming of the offer pdfs is now consistent based on a newly introduced naming class
